A bunch of later changes
- changes needed for PCZT integration
- splitting opcode.rs into multiple modules
- moving evaluation into `impl`s (interpreter.rs now holds mostly common
  types for interpretation and is depended on by opcode*.rs and script.rs)
- replace `Script` type with better-typed `Script<T>`, `ScriptPubKey`, and `ScriptSig<T>`
- add `enum AndMaybe<T, U>` for resolving push-only (should be its own module)
