+++ src/zcash_script.rs
@@ -1,9 +1,12 @@
 use std::num::TryFromIntError;
 
-use super::interpreter::*;
-use super::script::*;
-use super::script_error::*;
-use super::scriptnum::*;
+use secp256k1::ecdsa;
+
+use crate::external::pubkey::PubKey;
+use crate::interpreter::*;
+use crate::script::*;
+use crate::script_error::*;
+use crate::scriptnum::*;
 
 /// This maps to `zcash_script_error_t`, but most of those cases aren’t used any more. This only
 /// replicates the still-used cases, and then an `Unknown` bucket for anything else that might
@@ -58,25 +61,104 @@ pub trait ZcashScript {
     fn legacy_sigop_count_script(script: &[u8]) -> Result<u32, Error>;
 }
 
+/// All signature hashes are 32 bytes, since they are either:
+/// - a SHA-256 output (for v1 or v2 transactions).
+/// - a BLAKE2b-256 output (for v3 and above transactions).
+pub const SIGHASH_SIZE: usize = 32;
+
+/// A function which is called to obtain the sighash.
+///    - script_code: the scriptCode being validated. Note that this not always
+///      matches script_sig, i.e. for P2SH.
+///    - hash_type: the hash type being used.
+///
+/// The `extern "C"` function that calls this doesn’t give much opportunity for rich failure
+/// reporting, but returning `None` indicates _some_ failure to produce the desired hash.
+pub type SighashCalculator<'a> = &'a dyn Fn(&[u8], HashType) -> Option<[u8; SIGHASH_SIZE]>;
+
+pub struct CallbackTransactionSignatureChecker<'a> {
+    pub sighash: SighashCalculator<'a>,
+    pub lock_time: &'a ScriptNum,
+    pub is_final: bool,
+}
+
+impl CallbackTransactionSignatureChecker<'_> {
+    pub fn verify_signature(
+        sig: &ecdsa::Signature,
+        pubkey: &PubKey,
+        sighash: &[u8; SIGHASH_SIZE],
+    ) -> bool {
+        pubkey.verify(sighash, sig)
+    }
+}
+
+impl SignatureChecker for CallbackTransactionSignatureChecker<'_> {
+    fn check_sig(&self, sig: &Signature, vch_pub_key: &[u8], script_code: &[u8]) -> bool {
+        let pubkey = PubKey(vch_pub_key);
+
+        pubkey.is_valid()
+            && (self.sighash)(script_code, sig.sighash)
+                .map(|sighash| Self::verify_signature(&sig.sig, &pubkey, &sighash))
+                .unwrap_or(false)
+    }
+
+    fn check_lock_time(&self, lock_time: &ScriptNum) -> bool {
+        // There are two times of nLockTime: lock-by-blockheight
+        // and lock-by-blocktime, distinguished by whether
+        // nLockTime < LOCKTIME_THRESHOLD.
+        //
+        // We want to compare apples to apples, so fail the script
+        // unless the type of nLockTime being tested is the same as
+        // the nLockTime in the transaction.
+        if *self.lock_time < LOCKTIME_THRESHOLD && *lock_time >= LOCKTIME_THRESHOLD
+            || *self.lock_time >= LOCKTIME_THRESHOLD && *lock_time < LOCKTIME_THRESHOLD
+            // Now that we know we're comparing apples-to-apples, the
+            // comparison is a simple numeric one.
+            || lock_time > self.lock_time
+        {
+            false
+            // Finally the nLockTime feature can be disabled and thus
+            // CHECKLOCKTIMEVERIFY bypassed if every txin has been
+            // finalized by setting nSequence to maxint. The
+            // transaction would be allowed into the blockchain, making
+            // the opcode ineffective.
+            //
+            // Testing if this vin is not final is sufficient to
+            // prevent this condition. Alternatively we could test all
+            // inputs, but testing just this input minimizes the data
+            // required to prove correct CHECKLOCKTIMEVERIFY execution.
+        } else {
+            !self.is_final
+        }
+    }
+}
+
 /// A tag to indicate that the Rust implementation of zcash_script should be used.
 pub enum RustInterpreter {}
 
 impl RustInterpreter {
-    fn verify<T>(
+    fn verify<P>(
         script_pub_key: &[u8],
         script_sig: &[u8],
         flags: VerificationFlags,
-        payload: &mut T,
-        stepper: &impl Fn(&mut &[u8], &Script, &mut State, &mut T) -> Result<(), ScriptError>,
+        payload: &mut P,
+        stepper: &impl Fn(&dyn Evaluable, &[u8], &mut State, &mut P) -> Result<(), ScriptError>,
     ) -> Result<(), Error> {
-        verify_script(
-            &Script(script_sig),
-            &Script(script_pub_key),
-            flags,
-            payload,
-            stepper,
-        )
-        .map_err(|e| Error::Ok(Some(e)))
+        ScriptPubKey::from_bytes(script_pub_key)
+            .and_then(|(pub_key, _)| {
+                script_sigs_from_bytes(script_sig).and_then(|am| match am {
+                    AndMaybe::Only(sig) => {
+                        if flags.contains(VerificationFlags::SigPushOnly) {
+                            Err(ScriptError::SigPushOnly)
+                        } else {
+                            Script { sig, pub_key }.eval(flags, script_pub_key, payload, stepper)
+                        }
+                    }
+                    AndMaybe::Indeed(_, sig) => {
+                        Script { sig, pub_key }.eval(flags, script_pub_key, payload, stepper)
+                    }
+                })
+            })
+            .map_err(|e| Error::Ok(Some(e)))
     }
 }
 
@@ -98,12 +180,11 @@ impl StepResults {
 pub fn compare_step<'a>(
     flags: VerificationFlags,
     checker: &'a impl SignatureChecker,
-) -> impl Fn(&mut &[u8], &Script, &mut State, &mut StepResults) -> Result<(), ScriptError> + 'a {
+) -> impl Fn(&dyn Evaluable, &[u8], &mut State, &mut StepResults) -> Result<(), ScriptError> + 'a {
     move |pc, script, state, payload| {
-        let mut right_pc = *pc;
         let mut right_state = (*state).clone();
-        let left = eval_step(pc, script, flags, checker, state);
-        let right = eval_step(&mut right_pc, script, flags, checker, &mut right_state);
+        let left = pc.eval(flags, script, checker, state);
+        let right = pc.eval(flags, script, checker, &mut right_state);
 
         match (left, right) {
             (Ok(()), Ok(())) => {
@@ -138,8 +219,9 @@ impl ZcashScript for RustInterpreter {
     /// Returns the number of transparent signature operations in the
     /// transparent inputs and outputs of this transaction.
     fn legacy_sigop_count_script(script: &[u8]) -> Result<u32, Error> {
-        let cscript = Script(script);
-        Ok(cscript.get_sig_op_count(false))
+        ScriptPubKey::from_bytes(script)
+            .map_err(|e| Error::Ok(Some(e)))
+            .map(|cscript| cscript.0.get_sig_op_count(false))
     }
 
     fn verify_callback(
@@ -157,7 +239,7 @@ impl ZcashScript for RustInterpreter {
             lock_time: &lock_time_num,
             is_final,
         };
-        let stepper = eval_step2(flags, &checker);
+        let stepper = basic_evaluator(flags, &checker);
         Self::verify(script_pub_key, script_sig, flags, &mut (), &stepper)
     }
 }
@@ -185,42 +267,35 @@ pub mod testing {
 #[cfg(test)]
 mod tests {
     use super::{testing::*, *};
-    use crate::opcode::*;
     use crate::pattern::*;
     use hex::FromHex;
     use proptest::prelude::*;
 
     lazy_static::lazy_static! {
-        pub static ref SCRIPT_PUBKEY: Vec<u8> = Script::serialize(&pay_to_script_hash(&<[u8; 0x14]>::from_hex("c117756dcbe144a12a7c33a77cfa81aa5aeeb381").expect("valid hash")));
-        pub static ref SCRIPT_SIG: Vec<u8> = Script::serialize(&[
+        pub static ref SCRIPT_PUBKEY: Vec<u8> = ScriptPubKey(pay_to_script_hash(&<[u8; 0x14]>::from_hex("c117756dcbe144a12a7c33a77cfa81aa5aeeb381").expect("valid hash"))).to_bytes();
+        pub static ref SCRIPT_SIG: Vec<u8> = ScriptSig::new(vec![
             push_num(0),
             push_vec(&<[u8; 0x48]>::from_hex("3045022100d2ab3e6258fe244fa442cfb38f6cef9ac9a18c54e70b2f508e83fa87e20d040502200eead947521de943831d07a350e45af8e36c2166984a8636f0a8811ff03ed09401").expect("valid sig")),
             push_vec(&<[u8; 0x47]>::from_hex("3044022013e15d865010c257eef133064ef69a780b4bc7ebe6eda367504e806614f940c3022062fdbc8c2d049f91db2042d6c9771de6f1ef0b3b1fea76c1ab5542e44ed29ed801").expect("valid sig")),
-            push_script(&check_multisig(
+            push_script(&ScriptPubKey(check_multisig(
                 2,
                 &[
                     &<[u8; 0x21]>::from_hex("03b2cc71d23eb30020a4893982a1e2d352da0d20ee657fa02901c432758909ed8f").expect("valid key"),
                     &<[u8; 0x21]>::from_hex("029d1e9a9354c0d2aee9ffd0f0cea6c39bbf98c4066cf143115ba2279d0ba7dabe").expect("valid key"),
                     &<[u8; 0x21]>::from_hex("03e32096b63fd57f3308149d238dcbb24d8d28aad95c0e4e74e3e5e6a11b61bcc4").expect("valid key")
                 ],
-                false))
-        ].map(Opcode::PushValue));
+                false)))
+        ]).to_bytes();
     }
 
     fn sighash(_script_code: &[u8], _hash_type: HashType) -> Option<[u8; 32]> {
-        hex::decode("e8c7bdac77f6bb1f3aba2eaa1fada551a9c8b3b5ecd1ef86e6e58a5f1aab952c")
-            .unwrap()
-            .as_slice()
-            .first_chunk::<32>()
-            .copied()
+        <[u8; 32]>::from_hex("e8c7bdac77f6bb1f3aba2eaa1fada551a9c8b3b5ecd1ef86e6e58a5f1aab952c")
+            .ok()
     }
 
     fn invalid_sighash(_script_code: &[u8], _hash_type: HashType) -> Option<[u8; 32]> {
-        hex::decode("08c7bdac77f6bb1f3aba2eaa1fada551a9c8b3b5ecd1ef86e6e58a5f1aab952c")
-            .unwrap()
-            .as_slice()
-            .first_chunk::<32>()
-            .copied()
+        <[u8; 32]>::from_hex("08c7bdac77f6bb1f3aba2eaa1fada551a9c8b3b5ecd1ef86e6e58a5f1aab952c")
+            .ok()
     }
 
     fn missing_sighash(_script_code: &[u8], _hash_type: HashType) -> Option<[u8; 32]> {
